requirements.txt
setuptools~=68.0.0
imagehash~=4.3.1
numpy~=1.26.2
scenedetect~=0.6.2
pillow~=10.0.1